Extract CTF Input Data to Separate Structure
This commit creates a new helper structure,

    Opm::Connection::CTFProperties

that holds all of the individual items which go into calculating the
final connection transmissibility factor (CTF).  These include the
'Kh' product, the pressure equivalent radius ('r0'), the well-bore
radius ('rw'), and the skin factor.
the API of the constructor.  As an additional helper for the
implementation, and with a view towards simulation restart support,
we also store the value of the Peaceman formula denominator in
CTFProperties as

    double CTFProperties::peaceman_denom