Extract CTF Input Data to Separate Structure
Reimplement the Connection
constructor in terms of Opm::Connection::CTFProperties
and this changes the API of the constructor.

This, in turn, simplifies the implementation of CSKIN and obviates
the Connection::m_wpimult data member.  It also fixes a latent
problem in the current CSKIN implementation which would ignore any
explicitly input CTF values in favour of the pressure equivalent
radius.  By updating the explicit denominator with a new skin factor
instead we do not need to keep track of cumulative multiplier values
and we do not need to recompute the logarithmic term involving 'r0'
as part of CSKIN handling.